+++ src/main/java/kitchenpos/menu/domain/repository/MenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.menugroup.domain;
+package kitchenpos.menu.domain.repository;
 
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
