+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -8,12 +8,12 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.service.dto.MenuProductRequest;
 import kitchenpos.menu.service.dto.MenuRequest;
-import kitchenpos.menugroup.service.dto.MenuGroupResponse;
+import kitchenpos.menu.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.product.service.dto.ProductResponse;
 import kitchenpos.common.vo.Price;
 import kitchenpos.menu.service.MenuService;
-import kitchenpos.menugroup.service.MenuGroupService;
+import kitchenpos.menu.service.MenuGroupService;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
 import kitchenpos.supports.MenuFixture;
