+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -8,13 +8,13 @@ import java.util.List;
 import kitchenpos.order.service.dto.OrderStatusRequest;
 import kitchenpos.order.service.dto.TableGroupRequest;
 import kitchenpos.order.service.dto.TableRequest;
-import kitchenpos.menugroup.service.dto.MenuGroupResponse;
+import kitchenpos.menu.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.order.service.dto.OrderResponse;
 import kitchenpos.product.service.dto.ProductResponse;
 import kitchenpos.order.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
-import kitchenpos.menugroup.service.MenuGroupService;
+import kitchenpos.menu.service.MenuGroupService;
 import kitchenpos.order.service.OrderService;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.supports.IntegrationTest;
